requirements.txt
httpx>=0.26.0
rich>=13.7.0
newspaper4k>=0.9.0
